fixed issue in repository

+++ springboot_data_jpa/src/main/java/lk/harshana/dao/user/UserDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import lk.harshana.model.User;
 
-public interface UserDao extends CrudRepository<User, Long>{
+public interface UserDao extends CrudRepository<User, Integer>{
 
 	List<User> findByFirstName(String fName);
 	List<User> findByLastName(String lName);
